fix file naming/locations and bottom scrollbar
