+++ src/components/homepage/EmailForm.svelte
